Fix redirect loop in createGroupRoute mixin
When there was an invalid value selected, the GroupSelect
component would set the groupid back to null, and we would
forget the value and reroute back to the main page...

... BUT the remembered value watcher would see the value go back
to undefined, and, using the current route param (the invalid id)
set it all up again, hence an infinite loop.

The fix is to understand that once we've triggered a route change
we should do no more logic on this component. Once the route has
changed, it'll be all fresh again :)